big restructure with version bumps change the way the module works. Instead of wrapping the eks module we now take inputs that match the eks modue, enrich them, set them as outputs, and then can be passed to the eks module.
makes it easier to handle upgrades to the eks module as well as adds ability for caller to use whatever eks module functionality they want

all provisioning images have been updated to support 1.22 clusters

Removed calico in favor of new aws ip prefixes that support /28 ranges to allow for more ip/pods per nodes. For example a t3.medium used to only allow for 17 ips but now can handle 110. This is more than enough and removing calico greatly reduces the complexity and brittlness of the networking.